organizing files. added jsx and css file for the Instructors
