+++ frontend/main.jsx
@@ -1,5 +1,6 @@
-import App from "./App.jsx";
-import { Home } from "./Home.jsx";
+import App from "./pages/App.jsx";
+import { Home } from "./pages/Home.jsx";
+import { Instructors } from "./pages/Instructors.jsx";
 import "./index.css";
 import ReactDOM from "react-dom/client";
 import {
@@ -13,6 +14,7 @@ const router = createBrowserRouter(
   createRoutesFromElements(
     <Route path="/" element={<App />}>
       <Route path="/" element={<Home />} />
+      <Route path="/" element={<Instructors />} />
     </Route>
   )
 );
